+++ src/site/markdown/changelog.md
@@ -12,6 +12,7 @@
 ### Changed
 * Performed some code cleanups, moved some internal classes to different packages.
 * Made Searcher an abstract base class to SearcherImpl.
+* Added mock classes for Searcher, ForwardIndex, Terms; used them to add tests.
 
 ## Improvements up to v1.3.5